+++ draft-packages/stories/Popover.stories.tsx
@@ -2,6 +2,7 @@ import { DismissiblePositiveAutohide } from "@kaizen/component-library/stories/I
 import { Popover } from "@kaizen/draft-popover"
 import * as React from "react"
 import { Avatar } from "../avatar"
+import guidanceIcon from "@kaizen/component-library/icons/guidance.icon.svg"
 
 export default {
   title: "Popover (React)",
@@ -93,6 +94,21 @@ export const InformativeLargeWithSingleLine = () => (
 )
 
 InformativeLargeWithSingleLine.storyName = "Informative Large with singleLine"
+
+export const InformativeWithCustomIcon = () => (
+  <Container>
+    <Popover
+      heading="Informative"
+      variant="informative"
+      customIcon={guidanceIcon}
+    >
+      Popover body that explains something useful, is optional, and not critical
+      to completing a task.
+    </Popover>
+  </Container>
+)
+
+InformativeWithCustomIcon.storyName = "Informative with a custom icon"
 
 export const Positive = () => (
   <Container>