+++ packages/component-library/components/MenuList/MenuHeader.tsx
@@ -4,6 +4,9 @@ import * as React from "react"
 
 import styles from "./Menu.module.scss"
 
+/**
+ * @deprecated MenuHeader is deprecated. Please use draft-menu instead (it has its own menu header component).
+ */
 const MenuHeader = (props: { title: string }) => (
   <div className={styles.header}>
     <span className={styles.header__title}>{props.title}</span>