+++ draft-packages/select/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.11.13](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-select@1.11.12...@kaizen/draft-select@1.11.13) (2020-11-09)
+
+**Note:** Version bump only for package @kaizen/draft-select
+
+
+
+
+
+## [1.11.12](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-select@1.11.11...@kaizen/draft-select@1.11.12) (2020-11-09)
+
+**Note:** Version bump only for package @kaizen/draft-select
+
+
+
+
+
 ## [1.11.11](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-select@1.11.10...@kaizen/draft-select@1.11.11) (2020-10-21)
 
 **Note:** Version bump only for package @kaizen/draft-select